+++ src/jdom/jacdac-jdom.ts
@@ -35,6 +35,7 @@ export * from "./light"
 export * from "./registerserver"
 export * from "./serviceserver"
 export * from "./serviceprovider"
+export * from "./serverserviceprovider"
 export * from "./controlserver"
 export * from "./color"
 export * from "./bridge"
